fixed receiverThread to handle all bytes properly, added exact (platform independent) delay to send function

+++ src/driver_svh/SVHSerialInterface.h
@@ -105,7 +105,7 @@ private:
   //! serial device connected state
   bool m_connected;
 
-  int last_index;
+  uint8_t last_index;
 
   //! pointer to serial interface object
   boost::shared_ptr<Serial> m_serial_device;
